move dart extraction rules behind the indexer language registry
S3-DART (ARCH-012), one language-sized rollback unit following S3-KOTLIN.
Dart's tree-sitter rule file, compiled-rule cache, member-call callsite
marker, receiver-call resolution engine, and direct-call collector move out
of `codestory-indexer/src/lib.rs` into
`codestory-indexer/src/languages/dart.rs` as one `LanguageExtraction` row.
Every language-keyed dispatch that used to spell "dart" — parser config,
compiled rules, receiver-call collector, name-projection flags, comment
style, resolver selection, semantic family, callsite marker — now reads the
registry first and keeps its residual `match` for the fourteen languages
that have not moved; the last sibling package deletes those arms.

The move is mechanical. Diffing the 27 moved function bodies plus
`DartReceiverContext` against the lines deleted from `lib.rs`, ignoring
blank lines, is empty apart from three declared renames of moved entry
points: `collect_dart_receiver_call_edges` -> `receiver_call_specs`,
`collect_dart_direct_call_edges` -> `direct_call_specs`, and
`dart_member_call` -> `member_call`. 27 functions left `lib.rs`, 27 landed
in `languages/dart.rs` (28 counting the new `dart_language` constructor).

Three Dart surfaces deliberately stay in `lib.rs`, and all three are shared
seams rather than Dart content: `language_precise_call_specs` (Dart is the
only language with a precise-call collector, so `LanguageExtraction` has no
field for one — the residual arm now calls into the module),
`append_manual_receiver_call_edges`'s `language_name == "dart"`
placeholder-replacement branch (edge-assembly policy), and
`collect_shelf_route` in the framework-route scanner (the per-framework
collectors take non-uniform arguments, which is one change for all sixteen
languages). `LanguageRuleset::Dart` stays because the enum is the
compiled-rule cache key for every language at once.

Proof: `tests/language_extraction_snapshot.rs` gains a Dart row plus two
golden files pinning the full rendered projection of both Dart fixtures —
the fidelity_lab file through the store pipeline (resolution included) and
the tictactoe file through raw `index_file`. Both goldens were captured
before the production edit and are byte-identical after it (sha256
09dfbe75.. / 2c8f3c5b..); the Kotlin goldens are unchanged (a26ee4c6.. /
79df1ceb..), so the harness itself did not move. The threshold suites are
not sufficient on their own: changing Dart's qualified-name delimiter to
`::` leaves both `fidelity_regression` and `tictactoe_language_coverage`
green while the new snapshots go red.

Declared behavior changes:
- None to any indexing, resolution, or rendering output. Dart
  fidelity_regression and tictactoe_language_coverage projections are proven
  byte-identical before and after, and packet_search_eval is unmoved.
- `LanguageRuleset::Dart`'s arm in `compiled_rules` is now unreachable and
  returns a typed error instead of compiled rules; the registry answers
  first. Fail-closed by construction, matching the Kotlin arm.
- One additive row in `codestory_contracts::language_support::
  LANGUAGE_COMMENT_PROFILES` (`dart` -> `//`), and `dart` deleted from the
  CLI's residual comment roster in `output.rs::ansi_highlight_line`. The
  rendered marker is identical; `comment_markers_stay_identical_after_the_
  registry_lookup` proves it. The indexer's separate route-comment roster
  is untouched apart from deleting Dart's own name — the two rosters
  genuinely disagree and stay that way.
- `LANGUAGE_SUPPORT_PROFILES` is unchanged, so extension routing, claim
  tiers, `supported_extensions`, and `is_structural_source_path` are
  identical.
- No version surface touched; everything stays 0.16.3.
