+++ crates/codestory-indexer/src/semantic/mod.rs
@@ -413,9 +413,6 @@ fn dedicated_semantic_resolver(language: &str) -> Option<SemanticResolverKind> {
         "swift" => Some(SemanticResolverKind::Generic(GenericSemanticResolver::new(
             "swift",
         ))),
-        "dart" => Some(SemanticResolverKind::Generic(GenericSemanticResolver::new(
-            "dart",
-        ))),
         "bash" => Some(SemanticResolverKind::Generic(GenericSemanticResolver::new(
             "bash",
         ))),
@@ -601,7 +598,6 @@ fn language_family_bucket(language: &'static str) -> &'static str {
         "php" => "php",
         "csharp" => "csharp",
         "swift" => "swift",
-        "dart" => "dart",
         "bash" => "bash",
         language if is_structural_language_name(language) => "structural",
         _ => language,
