+++ crates/codestory-indexer/tests/fixtures/language_snapshots/dart_fidelity.txt
@@ -1,0 +1,35 @@
+node CLASS name=ConsoleNotifier qn=ConsoleNotifier canonical=<ROOT>/fidelity.dart:ConsoleNotifier line=7 col=1 end=11:2 file=FILE:<ROOT>/fidelity.dart@1
+node CLASS name=Event qn=Event canonical=<ROOT>/fidelity.dart:Event line=19 col=1 end=23:2 file=FILE:<ROOT>/fidelity.dart@1
+node CLASS name=Repository qn=Repository canonical=<ROOT>/fidelity.dart:Repository line=13 col=1 end=17:2 file=FILE:<ROOT>/fidelity.dart@1
+node CLASS name=Workflow qn=Workflow canonical=<ROOT>/fidelity.dart:Workflow line=25 col=1 end=35:2 file=FILE:<ROOT>/fidelity.dart@1
+node FILE name=<ROOT>/fidelity.dart qn=<ROOT>/fidelity.dart canonical=<ROOT>/fidelity.dart:<ROOT>/fidelity.dart:1 line=1 col=1 end=41:0 file=FILE:<ROOT>/fidelity.dart@1
+node FUNCTION name=orchestrateDart qn=orchestrateDart canonical=<ROOT>/fidelity.dart:orchestrateDart#0 line=37 col=1 end=37:23 file=FILE:<ROOT>/fidelity.dart@1
+node INTERFACE name=Notifier qn=Notifier canonical=<ROOT>/fidelity.dart:Notifier line=7 col=34 end=7:42 file=FILE:<ROOT>/fidelity.dart@1
+node METHOD name=ConsoleNotifier.notify qn=ConsoleNotifier.notify canonical=<ROOT>/fidelity.dart:ConsoleNotifier.notify#0 line=8 col=3 end=8:27 file=FILE:<ROOT>/fidelity.dart@1
+node METHOD name=Notifier.notify qn=Notifier.notify canonical=<ROOT>/fidelity.dart:Notifier.notify#0 line=4 col=3 end=4:28 file=FILE:<ROOT>/fidelity.dart@1
+node METHOD name=Repository.save qn=Repository.save canonical=<ROOT>/fidelity.dart:Repository.save#0 line=14 col=3 end=14:25 file=FILE:<ROOT>/fidelity.dart@1
+node METHOD name=Workflow.decorate qn=Workflow.decorate canonical=<ROOT>/fidelity.dart:Workflow.decorate#0 line=32 col=3 end=32:31 file=FILE:<ROOT>/fidelity.dart@1
+node METHOD name=Workflow.run qn=Workflow.run canonical=<ROOT>/fidelity.dart:Workflow.run#0 line=26 col=3 end=26:66 file=FILE:<ROOT>/fidelity.dart@1
+node MODULE name='dart:math' qn='dart:math' canonical=<ROOT>/fidelity.dart:'dart:math'#0 line=1 col=8 end=1:19 file=FILE:<ROOT>/fidelity.dart@1
+node UNKNOWN name=ConsoleNotifier qn=ConsoleNotifier canonical=<ROOT>/fidelity.dart:ConsoleNotifier#0 line=39 col=32 end=39:47 file=FILE:<ROOT>/fidelity.dart@1
+node UNKNOWN name=Event qn=Event canonical=<ROOT>/fidelity.dart:Event#0 line=39 col=16 end=39:21 file=FILE:<ROOT>/fidelity.dart@1
+node UNKNOWN name=Repository qn=Repository canonical=<ROOT>/fidelity.dart:Repository#0 line=39 col=51 end=39:61 file=FILE:<ROOT>/fidelity.dart@1
+node UNKNOWN name=Workflow qn=Workflow canonical=<ROOT>/fidelity.dart:Workflow#0 line=38 col=20 end=38:28 file=FILE:<ROOT>/fidelity.dart@1
+node UNKNOWN name=decorate qn=decorate canonical=<ROOT>/fidelity.dart:decorate#0 line=29 col=5 end=29:13 file=FILE:<ROOT>/fidelity.dart@1
+node UNKNOWN name=max qn=max canonical=<ROOT>/fidelity.dart:max#0 line=40 col=3 end=40:6 file=FILE:<ROOT>/fidelity.dart@1
+node UNKNOWN name=notify qn=notify canonical=<ROOT>/fidelity.dart:notify#0 line=27 col=14 end=27:20 file=FILE:<ROOT>/fidelity.dart@1
+node UNKNOWN name=print qn=print canonical=<ROOT>/fidelity.dart:print#0 line=9 col=5 end=9:10 file=FILE:<ROOT>/fidelity.dart@1
+node UNKNOWN name=print qn=print canonical=<ROOT>/fidelity.dart:print#1 line=15 col=5 end=15:10 file=FILE:<ROOT>/fidelity.dart@1
+node UNKNOWN name=run qn=run canonical=<ROOT>/fidelity.dart:run#0 line=39 col=12 end=39:15 file=FILE:<ROOT>/fidelity.dart@1
+node UNKNOWN name=save qn=save canonical=<ROOT>/fidelity.dart:save#0 line=28 col=16 end=28:20 file=FILE:<ROOT>/fidelity.dart@1
+edge CALL src=FUNCTION:orchestrateDart@37 dst=METHOD:Workflow.run@26 resolved_src=FUNCTION:orchestrateDart@37 resolved_dst=METHOD:Workflow.run@26 line=39 confidence=1.0000 certainty=Certain callsite=h0:39:1:h1 candidates=[]
+edge CALL src=METHOD:Workflow.run@26 dst=METHOD:Notifier.notify@4 resolved_src=METHOD:Workflow.run@26 resolved_dst=METHOD:Notifier.notify@4 line=27 confidence=1.0000 certainty=Certain callsite=h0:27:1:h2 candidates=[]
+edge CALL src=METHOD:Workflow.run@26 dst=METHOD:Repository.save@14 resolved_src=METHOD:Workflow.run@26 resolved_dst=METHOD:Repository.save@14 line=28 confidence=1.0000 certainty=Certain callsite=h0:28:1:h3 candidates=[]
+edge CALL src=METHOD:Workflow.run@26 dst=METHOD:Workflow.decorate@32 resolved_src=METHOD:Workflow.run@26 resolved_dst=METHOD:Workflow.decorate@32 line=29 confidence=1.0000 certainty=Certain callsite=h0:29:1:h4 candidates=[]
+edge IMPORT src=MODULE:'dart:math'@1 dst=MODULE:'dart:math'@1 resolved_src=MODULE:'dart:math'@1 resolved_dst=- line=0 confidence=- certainty=Certain callsite=- candidates=[]
+edge INHERITANCE src=CLASS:ConsoleNotifier@7 dst=INTERFACE:Notifier@7 resolved_src=- resolved_dst=- line=0 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=CLASS:ConsoleNotifier@7 dst=METHOD:ConsoleNotifier.notify@8 resolved_src=- resolved_dst=- line=0 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=CLASS:Repository@13 dst=METHOD:Repository.save@14 resolved_src=- resolved_dst=- line=0 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=CLASS:Workflow@25 dst=METHOD:Workflow.decorate@32 resolved_src=- resolved_dst=- line=0 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=CLASS:Workflow@25 dst=METHOD:Workflow.run@26 resolved_src=- resolved_dst=- line=0 confidence=- certainty=Certain callsite=- candidates=[]
+edge MEMBER src=INTERFACE:Notifier@7 dst=METHOD:Notifier.notify@4 resolved_src=- resolved_dst=- line=0 confidence=- certainty=Certain callsite=- candidates=[]
